% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play games !}
\usage{
play(game = "Yams")
}
\value{
launch a game of Yams or mind and return the sheets of score.
}
\description{
Play games !
}
\examples{
\dontrun{
play("Yams")
}
}
